requirements.txt
python-pptx>=0.6.21
bibtexparser>=1.2.0
openmeteo-requests>=1.2.0
requests-cache>=1.2.0
retry-requests>=2.0.0
requests>=2.31.0
lightgbm>=4.3.0

pandas
scikit-learn
matplotlib
ipykernel
numpy
xgboost